+++ src/operations/Completion.cpp
@@ -442,23 +442,27 @@ std::vector<lsp::CompletionItem> WorkspaceFolder::completion(const lsp::Completi
         }
 
         // Handle parentheses suggestions
-        if (entry.parens == Luau::ParenthesesRecommendation::CursorAfter)
+        if (config.completion.addParentheses)
         {
-            if (item.textEdit)
-                item.textEdit->newText += "()$0";
-            else
-                item.insertText = name + "()$0";
-            item.insertTextFormat = lsp::InsertTextFormat::Snippet;
-        }
-        else if (entry.parens == Luau::ParenthesesRecommendation::CursorInside)
-        {
-            if (item.textEdit)
-                item.textEdit->newText += "($1)$0";
-            else
-                item.insertText = name + "($1)$0";
-            item.insertTextFormat = lsp::InsertTextFormat::Snippet;
-            // Trigger Signature Help
-            item.command = lsp::Command{"Trigger Signature Help", "editor.action.triggerParameterHints"};
+            if (entry.parens == Luau::ParenthesesRecommendation::CursorAfter)
+            {
+                if (item.textEdit)
+                    item.textEdit->newText += "()$0";
+                else
+                    item.insertText = name + "()$0";
+                item.insertTextFormat = lsp::InsertTextFormat::Snippet;
+            }
+            else if (entry.parens == Luau::ParenthesesRecommendation::CursorInside)
+            {
+                auto parenthesesSnippet = config.completion.addTabstopAfterParentheses ? "($1)$0" : "($0)";
+                if (item.textEdit)
+                    item.textEdit->newText += parenthesesSnippet;
+                else
+                    item.insertText = name + parenthesesSnippet;
+                item.insertTextFormat = lsp::InsertTextFormat::Snippet;
+                // Trigger Signature Help
+                item.command = lsp::Command{"Trigger Signature Help", "editor.action.triggerParameterHints"};
+            }
         }
 
         if (entry.type.has_value())