+++ CHANGELOG.md
@@ -23,7 +23,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed reverse dependencies not being marked as dirty when using string requries due to unnormalised file paths
 - Fixed incorrect highlighting of unnamed parameters in signature help when multiple parameters present of same type
 - Fixed documentation not provided for some built-ins on hover
-- Fixed signature help highlighting of parameters named "\_"
+- Fixed signature help highlighting of parameters named `_`
+- Fixed documentation comments of parent function being attached to a nested function
 
 ## [1.12.1] - 2022-10-18